编辑器bug
