+++ mobile/.packages
@@ -1,4 +1,4 @@
-# Generated by pub on 2021-01-16 11:55:17.200790.
+# Generated by pub on 2021-01-17 11:34:03.144915.
 archive:file:///Users/virs/.fvm/versions/1.22.5-stable/.pub-cache/hosted/pub.flutter-io.cn/archive-2.0.13/lib/
 args:file:///Users/virs/.fvm/versions/1.22.5-stable/.pub-cache/hosted/pub.flutter-io.cn/args-1.6.0/lib/
 assets_audio_player:../packages/flutter_assetsAudioPlayer/lib/
