+++ mobile/lib/discuz.dart
@@ -18,7 +18,7 @@ import 'package:discuzq/api/forum.dart';
 import 'package:discuzq/router/routers.dart';
 import 'package:discuzq/views/exploreDelagate.dart';
 import 'package:discuzq/providers/appConfigProvider.dart';
-
+import 'package:discuzq/widgets/settings/privacyConfirm.dart';
 import 'widgets/ui/ui.dart';
 
 class Discuz extends StatefulWidget {
@@ -204,6 +204,8 @@ class __DiscuzAppDelegateState extends State<_DiscuzAppDelegate> {
   @override
   void initState() {
     this._getForumData();
+    Future.delayed(const Duration(seconds: 1))
+        .then((_) async => await _userPrivaciesNotice(context));
     super.initState();
   }
 
@@ -261,5 +263,21 @@ class __DiscuzAppDelegateState extends State<_DiscuzAppDelegate> {
     setState(() {
       _loaded = true;
     });
+  }
+
+  /// 弹出用户隐私提示
+  void _userPrivaciesNotice(BuildContext context) {
+    final dynamic appConf = context.read<AppConfigProvider>().appConf;
+    if (appConf['confrimedPrivacy'] != null &&
+        appConf['confrimedPrivacy'] == false) {
+      showModalBottomSheet(
+          context: context,
+          isDismissible: false,
+          enableDrag: false,
+          backgroundColor: Colors.transparent,
+          builder: (BuildContext context) {
+            return const PrivacyConfirm();
+          });
+    }
   }
 }
